+++ src/Schema/QuoterInterface.php
@@ -42,11 +42,10 @@ interface QuoterInterface
      * Splits full table name into parts.
      *
      * @param string $name The full name of the table.
-     * @param bool $withColumn Deprecated. Will be removed in version 2.0.0.
      *
      * @return string[] The table name parts.
      */
-    public function getTableNameParts(string $name, bool $withColumn = false): array;
+    public function getTableNameParts(string $name): array;
 
     /**
      * Ensures name is wrapped with `{{ and }}`.